# backend/openapi-spec-v2.yaml
openapi: 3.1.0
info:
  title: Govpulse Backend Service
  version: 2.0.0
  description: |
    This is the backend service for the Govpulse application, providing endpoints for user management, issue tracking, appointment scheduling, and more.
    With v2 it supports Clerk for user authentication and authorization.
...
paths:
...
  /api/v2/appointments/book-appointment:
    post:
      summary: Book an appointment
      operationId: bookAppointment
      description: Allows a user to book an appointment with government Authority.
      tags:
        - appointment-related
      requestBody:
        description: Contains the details of the appointment to be booked
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/models.bookAppointmentReqObj"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Appointment booked successfully
                  appointment_data:
                    $ref: "#/components/schemas/models.appointmentData"
        "400":
          description: Bad Request - Invalid input data
          content:
            text/plain:
              schema:
                type: string
                example: Bad Request - Invalid input data
        "401":
          description: Unauthorized - Invalid token or insufficient permissions
          content:
            text/plain:
              schema:
                type: string
                example: Unauthorized - Invalid token or insufficient permissions              
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error  
      deprecated: false
      security:
        - BearerAuth: []
...
  /api/v2/appointments/authority-appointments:
    get:
      summary: Get authority appointments
      operationId: getAuthorityAppointments
      description: Retrieves all appointments for the authenticated official's employed authority. Note that the authority is identified using the bearer token.
      tags:
        - appointment-related
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/models.appointmentData"
        "401":
          description: Unauthorized - Invalid token
          content:
            text/plain:
              schema:
                type: string
                example: Unauthorized - Invalid token
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error  
      deprecated: false
      security:
        - BearerAuth: []
  /api/v2/appointments/appointment-by-id/{appointment_id}:
    get:
      summary: Get appointment by ID
      operationId: getAppointmentById
      description: Retrieves a specific appointment by its ID
      tags:
        - appointment-related
      parameters:
        - name: appointment_id
          in: path
          required: true
          description: The ID of the appointment to retrieve
          schema:
            type: integer
            format: int64
            example: 12345
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.appointmentData"
        "404":
          description: Not Found - Appointment not found
          content:
            text/plain:
              schema:
                type: string
                example: Not Found - Appointment not found
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error  
      deprecated: false
      security: []
...
  /api/v2/appointments/attendees-of-appointment/{appointment_id}:
    get:
      summary: Get attendees of an appointment
      operationId: getAttendeesOfAppointment
      description: Retrieves all attendees of a specific appointment by its ID
      tags:
        - appointment-related
      parameters:
        - name: appointment_id
          in: path
          required: true
          description: The ID of the appointment whose attendees are to be retrieved
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/models.attendeeData"
        "404":
          description: Not Found - Appointment not found or no attendees available
          content:
            text/plain:
              schema:
                type: string
                example: Not Found - Appointment not found or no attendees available
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error  
      deprecated: false
      security: []
...
  /api/v2/appointments/remove-file-from-attachment:
    post:
      summary: Remove file from attachment
      operationId: removeFileFromAttachment
      description: Allows a user to remove a file from an appointment's attachment
      tags:
        - appointment-related
      requestBody:
        description: Contains the details of the file to be removed from the attachment
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                appointment_id:
                  type: integer
                  format: int64
                  example: 12345
                file_url:
                  type: string
                  format: uri
                  example: "https://example.com/attachment/file1.jpg"
              required:
                - appointment_id
                - file_url
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: File removed from attachment successfully
                  attachment_data:
                    $ref: "#/components/schemas/models.attachmentData"
        "400":
          description: Bad Request - Invalid input data
          content:
            text/plain:
              schema:
                type: string
                example: Bad Request - Invalid input data
        "401":
          description: Unauthorized - Invalid token or insufficient permissions
          content:
            text/plain:
              schema:
                type: string
                example: Unauthorized - Invalid token or insufficient permissions              
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error  
      deprecated: false
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    models.bookAppointmentReqObj:
      type: object
      properties:
        authority_id:
          type: integer
          format: int64
          example: 1
        date:
          type: string
          format: date-time
          example: "2025-08-15T00:00:00.000Z"
        time_slot:
          type: string
          example: "10:00 - 11:00"
        issue_id:
          type: integer
          format: int64
          nullable: true
          example: 12345
...
    models.appointmentData:
      type: object
      properties:
        appointment_id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
          example: 1
        authority_id:
          type: integer
          format: int64
          example: 2
        issue_id:
          type: integer
          format: int64
          nullable: true
          example: null
        date:
          type: string
          format: date-time
          example: "2025-08-15T00:00:00.000Z"
        time_slot:
          type: string
          example: "10:00 - 11:00"
        official_comment:
          type: string
          nullable: true
          example: null
...
    models.attendeeData:
      type: object
      properties:
        attendee_id:
          type: integer
          format: int64
          example: 1
        appointment_id:
          type: integer
          format: int64
          example: 1
        nic:
          type: string
          example: "425253456V"
        name:
          type: string
          example: "Jane Doe"
        phone_no:
          type: string
          format: phone
          example: "+94123456789"
...
    models.attachmentData:
      type: object
      properties:
        appointment_id:
          type: integer
          format: int64
          example: 1
        attachment_id:
          type: integer
          format: int64
          example: 1
        file_urls:
          type: array
          items:
            type: string
            format: uri
            example: "https://example.com/attachment.pdf"
      required:
        - appointment_id
        - file_urls
...